user CRUD operations fixed

+++ backend/routes/users.js
@@ -49,8 +49,10 @@ router.put("/login", async function (req, res, next) { //REVISAR
     if (err.message === "User not found") {
       return res.status(404).json({ error: "User not found" });
     }
-    console.error(`Error signing in: ${err}`);
-    return res.status(500).json({ error: "Internal server error" });
+    else{
+      console.error(`Error signing in: ${err}`);
+      return res.status(500).json({ error: "Internal server error" });
+    }
   }
 });
 
@@ -154,6 +156,13 @@ router.get("/getById", async function (req, res, next) {
 router.post("/create", async function (req, res, next) { 
   var newUser = req.body;
 
+  if(!req.session.user_id){
+    return res.status(401).json({ error: "User is not signed in" });
+  }
+  if(!req.session.role || req.session.role !== 'a'){
+    return res.status(403).json({ error: "You must be an admin to create a note" })
+  }
+
   if (!newUser || !newUser.userName || !newUser.email || !newUser.password) {
     res.status(400).send("Username, email and password are required to create a new user");
     return;
@@ -171,8 +180,12 @@ router.post("/create", async function (req, res, next) {
     });
 
   } catch (err) {
-    console.error(`Something went wrong trying to insert a document: ${err}\n`);
-    res.status(500).send("Internal server error");
+    if(err.message === "Email already in use"){
+      return res.status(400).send("Email already in use")
+    }else{
+      console.error(`Something went wrong trying to insert a document: ${err}\n`);
+      res.status(500).send("Internal server error");
+    }
   }
 });
 
@@ -188,6 +201,12 @@ router.put("/edit", async function (req, res, next) {
       .status(400)
       .send("_id is required to update a user");
     return;
+  }
+  if(!req.session.user_id){
+    return res.status(401).send("User is not signed in")
+  }
+  if(!req.session.role || req.session.role !== 'a'){
+    return res.status(403).send("User is not signed in")
   }
 
   try{
@@ -213,12 +232,17 @@ router.delete("/delete", async function (req, res, next) {
      .send("ID is required to delete a user");
    return;
  }
+ if(!req.session.user_id){
+  return res.status(401).send("User is not signed in")
+ }
+ if(req.session.user_id === user.id || req.session.role === 'a'){
   try{
     await userModel.deleteUserById(user.id);
     console.log("User deleted successfully");
     res.status(200).json({
       message: "User deleted successfully"
     });
+    return;
   }catch (err) {
     if(err.message === "User not found"){
       res.status(404).send("User not found");
@@ -226,7 +250,11 @@ router.delete("/delete", async function (req, res, next) {
     }else{
       console.error(`Something went wrong trying to delete a document: ${err}\n`);
       res.status(500).send("Internal server error");
+      return
     }
+  }
+  }else{
+    return res.status(403).send("You must be admin or the selected user to delete it")
   }
 });
